Fix problem with defined min/max macros if windows.h is included
If cvui is used on Windows and windows.h is included, or any other lib that includes it, e.g. https://github.com/tronkko/dirent, then min() and max() will be defined. It will cause all sorts of problems.

This commit fix that problem by introducing a check for the definition of those macros. If they are found, they are saved, undefined, then re-defined at the end of cvui.h. This should fix problems with windows.h and make cvui.h play nice with other libs on Windows. Hopefully.