Organize examples involving movable/minimizable window
I've created a new set of examples: the ui-*. They are organized in a way to be modular and easy for others to use code related to creating more complex UI based on cvui components. The very first two examples in the ui-* category are ui-enhanced-canny and ui-enhanced-window-component. Both contain contributions from @shengyu7697 and @abreheret. Thank you very much, guys! I hope you like the re-organization I conducted on the files you submitted.